pyproject.toml
[tool.poetry]
name = "spade-cli"
version = "0.2.4"
description = "You can't dig without a spade! - a super-powered networking tool!"
authors = ["Vivaan Verma <hello@vivaanverma.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "spade" }
]

[tool.poetry.scripts]
spade = "spade.cli:spade"

[tool.poetry.dependencies]
python = "^3.10"
typer = { extras = ["all"], version = "^0.4.0"}
rich = "^10.14"
dnspython = "^2.0"
pydantic = "^1.8"
icmplib = "^3.0.2"
ipwhois = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
